Move `drawProgressBar` function to it's own file
