+++ appOPHD/UI/ProgressBar.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <NAS2D/Math/Rectangle.h>
+
+
+void drawProgressBar(int value, int max, NAS2D::Rectangle<int> rect, int padding = 4);
